requirements.txt
crewai>=0.28.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
groq>=0.4.0
openai>=1.12.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
scipy>=1.11.0
Pillow>=10.0.0
